Backend Completo

+++ backend/controllers/tareasControllers.js
@@ -3,7 +3,7 @@ const Tarea = require('../models/tareasModel')
 
 const getTareas = asyncHandler(async (req, res) => {
 
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({ user: req.user.id })
 
     res.status(200).json(tareas)
 })
@@ -17,6 +17,7 @@ const setTarea = asyncHandler(async (req, res) => {
 
     const tarea = await Tarea.create({
         texto: req.body.texto,
+        user: req.user.id
     })
     res.status(201).json(tarea)
 })
@@ -30,8 +31,14 @@ const updateTarea = asyncHandler(async (req, res) => {
         throw new Error('La tarea no fué encontrada')
     }
 
-    const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
-    res.status(200).json(updatedTarea)
+    //verificar que la tarea pertenezca al usuario logeado
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
+        const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+        res.status(200).json(updatedTarea)
+    }
 })
 
 const deleteTarea = asyncHandler(async (req, res) => {
@@ -43,10 +50,16 @@ const deleteTarea = asyncHandler(async (req, res) => {
         throw new Error('La tarea no fué encontrada')
     }
 
-    tarea.deleteOne()
-    //const deletedTarea = await Tarea.findByIdAndDelete(req.params.id)
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
+        tarea.deleteOne()
+        //const deletedTarea = await Tarea.findByIdAndDelete(req.params.id)
 
-    res.status(200).json({ id: tarea._id })
+        res.status(200).json({ id: tarea._id })
+    }
+
 })
 
 module.exports = {
